refactor: 리뷰 통계 topKeywords 타입을 Integer로 변경하여 일관성 유지

+++ src/main/java/com/byunsum/ticket_reservation/admin/dto/ReviewStatsResponse.java
@@ -11,5 +11,5 @@ public record ReviewStatsResponse(@Schema(description = "리뷰 총 개수")
                                   Map<String, Long> sentimentCount,
 
                                   @Schema(description = "자주 언급된 키워드")
-                                  Map<String, Long> topKeywords) {
+                                  Map<String, Integer> topKeywords) {
 }
